Perform Bayesian finemapping using CARMA

This function is a wrapper around CARMA::CARMA() that takes a dataframe containing variants at a locus and performs Bayesian finemapping based on their signed z-scores and an LD reference panel. The function requires a plink-formatted LD reference panel (bfile), which will be used to generate a signed LD matrix at the locus. CARMA was described in Yang et al. (Nature Genetics 2023; https://doi.org/10.1038/s41588-023-01392-0)

Usage

run_carma(
  df,
  snp_col,
  z_col,
  effect_allele_col,
  outlier_switch = TRUE,
  bfile,
  threads = 1,
  memory = 16000,
  plink_bin,
  ...
)

Arguments

df

Dataframe containing variants at a locus for finemapping

snp_col

Name of column containing SNP identifiers

z_col

Name of column containing signed Z-scores (relative to effect allele)

effect_allele_col

Name of column containing effect alleles

outlier_switch

Whether to remove outliers when performing finemapping. Argument is passed to the outlier.switch argument of CARMA::CARMA()

bfile

Path to plink bfile of reference panel that will be used to extract LD

threads

Number of threads (default = 1)

memory

Memory limit (default = 16000 MB)

plink_bin

Path to plink executable

...

additional arguments passsed to CARMA::CARMA()

Value

A dataframe containing the input dataframe, and additional columns denoting which credible set (CS) each variant belongs to, as well as the posterior inclusion probability (PIP), and an ld_error column noting whether there were problems generating the LD matrix that limited fine-mapping.

See also

Examples

if (FALSE) { # \dontrun{
run_carma(locus_df, snp_col = SNP, z_col = z, effect_allele_col = allele1)
} # }